+++ PocketCartApp.Domain/Domain Models/Product.cs
@@ -10,21 +10,21 @@ namespace PocketCartApp.Domain.Domain_Models
     public class Product : BaseEntity
     {
         [Required]
-        public string? ProductName { get; set; }
+        public required string ProductName { get; set; }
         [Required]
         public double ProductPrice { get; set; }
         [Required]
-        public Guid? CategoryId { get; set; }
+        public Guid CategoryId { get; set; }
         public string? CategoryName { get; set; }
         [Required]
         public DateOnly ExpirationDate { get; set; }
         [Required]
-        public double? quantity { get; set; }
+        public double quantity { get; set; }
         [Required]
-        public Guid? ManufacturerId { get; set; }
+        public Guid ManufacturerId { get; set; }
         public Manufacturer? Manufacturer { get; set; }
         public virtual ICollection<ShoppingCart>? AllShoppingCarts { get; set; }
-        public string? Barcode { get; set; }
-        public string? BarcodeImagePath { get; set; }
+        public required string Barcode { get; set; }
+        public required string BarcodeImagePath { get; set; }
     }
 }